requirements.txt
absl-py==0.9.0
alembic==1.4.2
apache-airflow==1.10.10
apispec==1.3.3
appnope==0.1.0
argcomplete==1.11.1
astunparse==1.6.3
attrs==19.3.0
authlib==0.14.3
babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.5
blis==0.4.1
cached-property==1.5.1
cachetools==4.1.0
catalogue==1.0.0
cattrs==0.9.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.31
cryptography==2.9.2
crython==0.2.0
cycler==0.10.0
cymem==2.0.3
cython==0.29.19
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
dnspython==1.16.0
docutils==0.16
email-validator==1.1.1
entrypoints==0.3
flask-admin==1.5.4
flask-appbuilder==2.3.4; python_version >= '3.6'
flask-babel==1.0.0
flask-caching==1.3.3
flask-jwt-extended==3.24.1
flask-login==0.4.1
flask-openid==1.2.5
flask-sqlalchemy==2.4.1
flask-swagger==0.2.13
flask-wtf==0.14.3
flask==1.1.2
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
google-auth-oauthlib==0.4.1
google-auth==1.15.0
google-pasta==0.2.0
graphviz==0.14
grpcio==1.29.0
gunicorn==19.10.0
h5py==2.10.0
idna==2.9
ipykernel==5.3.0
ipython-genutils==0.2.0
ipython==7.14.0
ipywidgets==7.5.1
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.17.0
jinja2==2.10.3
joblib==0.15.1
json-merge-patch==0.2
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter==1.0.0
jupyterlab-server==1.1.4
jupyterlab==2.1.2
keras-applications==1.0.8
keras-preprocessing==1.1.2
keras==2.3.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lemmy==2.1.0
lockfile==0.12.2
mako==1.1.2
markdown==2.6.11
markupsafe==1.1.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.0
marshmallow==2.21.0
matplotlib==3.2.1
mistune==0.8.4
more-itertools==8.3.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.6
nltk==3.5
notebook==6.0.3
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.7.0
pendulum==1.4.4
pexpect==4.8.0
pickleshare==0.7.5
pillow==7.1.2
plac==1.1.3
preshed==3.0.2
prison==0.1.3
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.1
psutil==5.7.0
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pygments==2.6.1
pyjwt==1.7.1
pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.16.0
pysqlite3==0.4.3
pytesseract==0.3.4
python-daemon==2.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2020.1
pytzdata==2019.3
pyyaml==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
qtpy==1.9.0
regex==2020.5.14
requests-oauthlib==1.3.0
requests==2.23.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
send2trash==1.5.0
setproctitle==1.1.10
simple-salesforce==1.0.0
simplecv==1.3
six==1.14.0
spacy==2.2.4
sqlalchemy-jsonfield==0.9.0; python_version >= '3.5'
sqlalchemy-utils==0.36.5
sqlalchemy==1.3.17
srsly==1.0.2
tabulate==0.8.7
tenacity==4.12.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.2
thinc==7.4.0
threadpoolctl==2.0.0
thrift==0.13.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
treelib==1.6.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.9
waitress==1.4.3
wasabi==0.6.0
wcwidth==0.1.9
webencodings==0.5.1
werkzeug==0.16.1
wheel==0.34.2; python_version >= '3'
widgetsnbextension==3.5.1
wrapt==1.12.1
wtforms==2.3.1
xgboost==1.1.0
zope.deprecation==4.4.0
torch==1.5.0
torchvision==0.6.0
